Discord-bot/requirements.txt
discord.py
asyncio
cerberus